pyproject.toml
[tool.poetry]
name = "nonebot-plugin-pokesomeone"
version = "0.1.0"
description = "指定戳一戳"
authors = ["canxin <1969730106@qq.com>"]
readme = "README.md"
packages = [{ include = "nonebot_plugin_pokesomeone" }]

[tool.poetry.dependencies]
python = "^3.8"
nonebot-adapter-onebot = "^2.2.3"
nonebot2 = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
